test: migrate from Jest to Rstest
