--- jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  displayName: 'CspHtmlRspackPlugin',
-  roots: ['<rootDir>'],
-  testMatch: ['**/*.jest.js'],
-  testPathIgnorePatterns: ['/node_modules/'],
-  clearMocks: true,
-};
