+++ package.json
@@ -7,15 +7,14 @@
   "scripts": {
     "eslint": "eslint .",
     "eslint:fix": "eslint . --fix",
-    "jest": "NODE_OPTIONS=--experimental-vm-modules jest --config=./jest.config.js",
+    "test:unit": "rstest",
     "test:real-content-hash": "node ./test-real-content-hash.mjs",
-    "jest:watch": "NODE_OPTIONS=--experimental-vm-modules jest --watch --verbose=false --config=./jest.config.js",
-    "jest:coverage:generate": "NODE_OPTIONS=--experimental-vm-modules jest --coverage --config=./jest.config.js",
-    "jest:coverage:clean": "rimraf  ./coverage",
-    "jest:coverage:upload": "npx codecov",
-    "jest:coverage": "npm run jest:coverage:clean && npm run jest:coverage:generate && npm run jest:coverage:upload",
-    "test": "npm run eslint && npm run jest && npm run test:real-content-hash",
-    "test:coverage": "npm run test && npm run jest:coverage",
+    "test:watch": "rstest --watch",
+    "test:coverage:generate": "rstest --coverage",
+    "test:coverage:clean": "rimraf ./coverage",
+    "test:coverage:upload": "npx codecov",
+    "test": "npm run eslint && npm run test:unit && npm run test:real-content-hash",
+    "test:coverage": "npm run test && npm run test:coverage:clean && npm run test:coverage:generate && npm run test:coverage:upload",
     "release": "node ./scripts/release.mjs"
   },
   "homepage": "https://github.com/rspack-contrib/csp-rspack-plugin",
@@ -54,7 +53,8 @@
   },
   "devDependencies": {
     "@continuous-auth/client": "2.3.2",
-    "babel-jest": "^30.4.1",
+    "@rstest/core": "0.11.4",
+    "@rstest/coverage-istanbul": "0.11.4",
     "codecov": "^3.8.1",
     "css-loader": "^6.5.1",
     "eslint": "^8.5.0",
@@ -63,7 +63,6 @@
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-prettier": "^4.0.0",
     "html-webpack-plugin": "^5.0.0-alpha.15",
-    "jest": "^30.4.2",
     "memory-fs": "^0.5.0",
     "mini-css-extract-plugin": "^2.4.5",
     "prettier": "^2.2.1",
